fix: use `fork` properly
